pyproject.toml
[tool.poetry]
name = "currency-converter"
version = "0.1.0"
description = ""
authors = ["Anton Gromov <seed14112@yaandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
drf-yasg = "^1.21.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
